fix(core): made changes and resolved some code smells
