+++ libs/core/src/main/java/com/sourcefuse/jarc/core/awares/ApplicationAwareBeanUtils.java
@@ -1,22 +1,28 @@
 package com.sourcefuse.jarc.core.awares;
 
 import jakarta.persistence.EntityManager;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ApplicationAwareBeanUtils {
+public class ApplicationAwareBeanUtils implements ApplicationContextAware {
 
-  @Autowired
-  private ApplicationContext applicationContext;
+  private ApplicationContext context;
+
+  @Override
+  public void setApplicationContext(ApplicationContext applicationContext)
+    throws BeansException {
+    context = applicationContext;
+  }
 
   public <T> T getBean(Class<T> beanClass) {
-    return applicationContext.getBean(beanClass);
+    return context.getBean(beanClass);
   }
 
   public EntityManager getNewEntityManager() {
-    EntityManager em = applicationContext.getBean(EntityManager.class);
+    EntityManager em = context.getBean(EntityManager.class);
     return em.getEntityManagerFactory().createEntityManager();
   }
 }
