+++ libs/core/src/main/java/com/sourcefuse/jarc/core/constants/SoftDeleteRepositoryConstants.java
@@ -2,10 +2,6 @@ package com.sourcefuse.jarc.core.constants;
 
 public final class SoftDeleteRepositoryConstants {
 
-  private SoftDeleteRepositoryConstants() {
-    throw new IllegalStateException("Utility class");
-  }
-
   public static final String DELETED_FIELD = "deleted";
   public static final String ID_MUST_NOT_BE_NULL =
     "The given id must not be null!";
@@ -15,4 +11,8 @@ public final class SoftDeleteRepositoryConstants {
     "The entity must not be null!";
   public static final String ENTITIES_MUST_NOT_BE_NULL =
     "Entities must not be null!";
+
+  private SoftDeleteRepositoryConstants() {
+    throw new IllegalStateException("Utility class");
+  }
 }
